requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-multipart
aiofiles
httpx
asyncpg
sqlalchemy
sqlmodel
alembic

# YouTube & Media
yt-dlp
opencv-python-headless
Pillow
imagehash

# Vector DB
qdrant-client

# ML & AI
numpy
torch
torchvision
transformers
sentence-transformers
clip-retrieval
timm
google-genai

# Utils
python-dotenv
structlog
tenacity
